requirements.txt
PyTelegramBotApi
python-dotenv